Refactor part 1 to use the approach from part 2
Originally, I tried to use a linked list to solve this problem,
because the problem called for lots of random access insertions. This
ended up being *way* too slow for part two however...

So I switched to another approach for part two that uses a "hash" map
instead, and the program is orders of magnitude faster.

In this commit, I refactored part one to use the same code as part
two and fixed the memory leak in part two.